requirements.txt
Flask>=2.0
psutil>=5.9
requests>=2.25
python-dotenv>=1.0
chardet>=5.0
eventlet>=0.33